Move header and fetching logic to Playbook
